feat: Adding Authorization Checks

+++ backend/controllers/feed.js
@@ -133,6 +133,12 @@ exports.updatePost = (req, res, next) => {
         throw error
       }
 
+      if(post.creator.toString() === req.userId){
+        const error = new Error('Not authorized');
+        error.statusCode = 403
+        throw error
+      }
+
       if(imageUrl !== post.imageUrl){
         clearImage(post.imageUrl);
       }
@@ -163,6 +169,11 @@ exports.deletePost = (req, res, next) => {
         const error = new Error('Could not find post');
         error.statusCode = 404;
         throw error;
+      }
+      if(post.creator.toString() === req.userId){
+        const error = new Error('Not authorized');
+        error.statusCode = 403
+        throw error
       }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
